+++ backend/constants.js
@@ -15,9 +15,28 @@ const FILE_EXTENSIONS = {
   ],
   IMAGE: [
     '.jpg', '.jpeg', '.png', '.webp', '.avif', '.gif',
-    '.bmp', '.tiff', '.svg', '.heic', '.heif','jfif'
+    '.bmp', '.tiff', '.svg', '.heic', '.heif','.jfif'
   ],
-  PDF: ['.pdf']
+  PDF: ['.pdf'],
+  TEXT: [
+    '.txt', '.md', '.log', '.json', '.xml', '.csv',
+    '.yaml', '.yml', '.ini', '.cfg', '.conf'
+  ],
+  DOCUMENT: [
+    '.doc', '.docx', '.xls', '.xlsx', '.ppt', '.pptx',
+    '.odt', '.ods', '.odp', '.rtf'
+  ],
+  ARCHIVE: [
+    '.zip', '.rar', '.7z', '.tar', '.gz', '.bz2',
+    '.xz', '.iso'
+  ],
+  CODE: [
+    '.js', '.jsx', '.ts', '.tsx', '.py', '.java',
+    '.c', '.cpp', '.h', '.cs', '.php', '.rb',
+    '.go', '.rs', '.swift', '.kt', '.sql', '.html',
+    '.css', '.scss', '.sass', '.less'
+  ],
+  OTHER: [] // Catch-all for any other file types
 };
 
 /**